Consolidate Orders Compose UI ownership
