+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/list/OrderListOrderRow.kt
@@ -1,8 +1,14 @@
 package com.woocommerce.android.ui.orders.list
 
+import androidx.compose.animation.core.animate
+import androidx.compose.animation.core.tween
 import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.MutatePriority
 import androidx.compose.foundation.background
 import androidx.compose.foundation.combinedClickable
+import androidx.compose.foundation.gestures.Orientation
+import androidx.compose.foundation.gestures.draggable
+import androidx.compose.foundation.gestures.rememberDraggableState
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -10,6 +16,8 @@ import androidx.compose.foundation.layout.ExperimentalLayoutApi
 import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.absoluteOffset
+import androidx.compose.foundation.layout.absolutePadding
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.heightIn
@@ -20,27 +28,31 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableFloatStateOf
+import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.AbsoluteAlignment
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.focus.focusRequester
-import androidx.compose.ui.input.key.Key
-import androidx.compose.ui.input.key.KeyEventType
-import androidx.compose.ui.input.key.key
-import androidx.compose.ui.input.key.onPreviewKeyEvent
-import androidx.compose.ui.input.key.type
+import androidx.compose.ui.layout.onSizeChanged
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.semantics.CustomAccessibilityAction
 import androidx.compose.ui.semantics.Role
+import androidx.compose.ui.semantics.clearAndSetSemantics
 import androidx.compose.ui.semantics.customActions
 import androidx.compose.ui.semantics.requestFocus
 import androidx.compose.ui.semantics.role
 import androidx.compose.ui.semantics.selected
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
 import com.woocommerce.android.R
 import com.woocommerce.android.ui.compose.designsystem.WooTheme
@@ -48,6 +60,7 @@ import com.woocommerce.android.ui.compose.designsystem.component.WooBadge
 import com.woocommerce.android.ui.compose.designsystem.component.WooBadgeDefaults
 import com.woocommerce.android.ui.compose.designsystem.icons.CheckSmall
 import com.woocommerce.android.ui.compose.designsystem.icons.WooIcons
+import kotlin.math.roundToInt
 
 @OptIn(ExperimentalFoundationApi::class, ExperimentalLayoutApi::class)
 @Composable
@@ -93,17 +106,6 @@ internal fun OrderListOrderRow(
                 .testTag(OrderListTestTags.orderRow(order.orderId))
                 .background(background)
                 .focusRequester(focusRequester)
-                .onPreviewKeyEvent { event ->
-                    if (event.key != Key.Spacebar) return@onPreviewKeyEvent false
-                    when (event.type) {
-                        KeyEventType.KeyDown -> true
-                        KeyEventType.KeyUp -> {
-                            onSelectionToggle()
-                            true
-                        }
-                        else -> false
-                    }
-                }
                 .combinedClickable(
                     onClick = activate,
                     onLongClick = onLongPress,
@@ -138,6 +140,130 @@ internal fun OrderListOrderRow(
                 order = order,
                 modifier = Modifier.weight(1f),
             )
+        }
+    }
+}
+
+@Composable
+private fun OrderListSwipeToComplete(
+    orderId: Long,
+    isCompleted: Boolean,
+    isEnabled: Boolean,
+    onMarkCompleted: () -> Unit,
+    canHandleDelta: () -> Boolean = { true },
+    canCommit: () -> Boolean = { true },
+    modifier: Modifier = Modifier,
+    content: @Composable (Modifier) -> Unit,
+) {
+    var widthPx by remember(orderId) { mutableIntStateOf(0) }
+    var offsetPx by remember(orderId) { mutableFloatStateOf(0f) }
+    val dragState = rememberDraggableState { delta ->
+        if (widthPx == 0 || !canHandleDelta()) return@rememberDraggableState
+
+        val maximumTravel = maximumOrderSwipeTravel(widthPx, isCompleted)
+        offsetPx = (offsetPx + delta).coerceIn(
+            minimumValue = -maximumTravel,
+            maximumValue = 0f,
+        )
+    }
+
+    LaunchedEffect(isEnabled, orderId) {
+        if (!isEnabled) {
+            dragState.drag(MutatePriority.PreventUserInput) {
+                offsetPx = 0f
+            }
+        }
+    }
+
+    Box(
+        modifier = modifier
+            .onSizeChanged { widthPx = it.width }
+            .draggable(
+                state = dragState,
+                orientation = Orientation.Horizontal,
+                enabled = isEnabled,
+                startDragImmediately = false,
+                reverseDirection = false,
+                onDragStopped = {
+                    val shouldMarkCompleted = canCommit() &&
+                        !isCompleted &&
+                        shouldCompleteOrderSwipe(offsetPx, widthPx)
+                    if (shouldMarkCompleted) {
+                        onMarkCompleted()
+                    }
+                    dragState.drag {
+                        animate(
+                            initialValue = offsetPx,
+                            targetValue = 0f,
+                            animationSpec = tween(),
+                        ) { value, _ ->
+                            offsetPx = value
+                        }
+                    }
+                },
+            ),
+    ) {
+        OrderListSwipeBackground(
+            orderId = orderId,
+            isCompleted = isCompleted,
+            showAction = isEnabled,
+            modifier = Modifier.matchParentSize(),
+        )
+        content(
+            Modifier.absoluteOffset {
+                IntOffset(offsetPx.roundToInt(), 0)
+            }
+        )
+    }
+}
+
+@Composable
+private fun OrderListSwipeBackground(
+    orderId: Long,
+    isCompleted: Boolean,
+    showAction: Boolean,
+    modifier: Modifier = Modifier,
+) {
+    val background = if (isCompleted) {
+        colorResource(R.color.color_on_surface_disabled)
+    } else {
+        WooTheme.colors.primary
+    }
+    Box(
+        modifier = modifier
+            .background(background)
+            .then(
+                if (!isCompleted && showAction) {
+                    Modifier.testTag(OrderListTestTags.swipeReveal(orderId))
+                } else {
+                    Modifier
+                }
+            ),
+    ) {
+        if (!isCompleted && showAction) {
+            Column(
+                modifier = Modifier
+                    .align(AbsoluteAlignment.TopRight)
+                    .absolutePadding(
+                        top = WooTheme.padding.padding5,
+                        right = WooTheme.padding.padding5,
+                    )
+                    .clearAndSetSemantics { },
+                horizontalAlignment = Alignment.CenterHorizontally,
+                verticalArrangement = Arrangement.spacedBy(WooTheme.spacing.space2),
+            ) {
+                Icon(
+                    imageVector = WooIcons.Regular.CheckSmall,
+                    contentDescription = null,
+                    tint = WooTheme.colors.onPrimary,
+                    modifier = Modifier.size(WooTheme.iconSize.size24),
+                )
+                Text(
+                    text = stringResource(R.string.orderlist_mark_completed),
+                    color = WooTheme.colors.onPrimary,
+                    style = WooTheme.text.bodyLarge.regular,
+                )
+            }
         }
     }
 }
@@ -242,3 +368,15 @@ private fun OrderContent(
 private val ORDER_ROW_MIN_HEIGHT = 96.dp
 private val SELECTION_SLOT_START_INSET = 12.dp
 private val SELECTION_INDICATOR_SIZE = 40.dp
+private const val COMPLETED_MAXIMUM_TRAVEL = 0.1f
+private const val COMPLETION_THRESHOLD = 0.5f
+private const val FULL_TRAVEL = 1f
+
+internal fun shouldCompleteOrderSwipe(offsetPx: Float, widthPx: Int): Boolean {
+    return widthPx > 0 && offsetPx < -(widthPx * COMPLETION_THRESHOLD)
+}
+
+internal fun maximumOrderSwipeTravel(widthPx: Int, isCompleted: Boolean): Float {
+    val travelFraction = if (isCompleted) COMPLETED_MAXIMUM_TRAVEL else FULL_TRAVEL
+    return widthPx * travelFraction
+}
