+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/list/OrderListScreen.kt
@@ -1,25 +1,51 @@
+@file:Suppress("MagicNumber")
+
 package com.woocommerce.android.ui.orders.list
 
+import androidx.annotation.StringRes
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
 import androidx.compose.animation.scaleIn
 import androidx.compose.animation.scaleOut
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.ExperimentalLayoutApi
+import androidx.compose.foundation.layout.FlowRow
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.selection.toggleable
 import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.semantics.Role
+import androidx.compose.ui.semantics.heading
+import androidx.compose.ui.semantics.semantics
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.tooling.preview.PreviewLightDark
 import com.woocommerce.android.R
 import com.woocommerce.android.ui.compose.designsystem.WooTheme
+import com.woocommerce.android.ui.compose.designsystem.component.WooButtonSize
+import com.woocommerce.android.ui.compose.designsystem.component.WooDivider
+import com.woocommerce.android.ui.compose.designsystem.component.WooNoticeBanner
+import com.woocommerce.android.ui.compose.designsystem.component.WooNoticeBannerTone
+import com.woocommerce.android.ui.compose.designsystem.component.WooOutlinedButton
+import com.woocommerce.android.ui.compose.designsystem.foundation.WooDesignSystemThemeWithBackground
+import com.woocommerce.android.ui.compose.designsystem.icons.AngleDown
+import com.woocommerce.android.ui.compose.designsystem.icons.AngleUp
+import com.woocommerce.android.ui.compose.designsystem.icons.CircleInfo
 import com.woocommerce.android.ui.compose.designsystem.icons.Plus
 import com.woocommerce.android.ui.compose.designsystem.icons.WooIcons
 
@@ -123,3 +149,326 @@ internal fun OrderListScreen(
         }
     }
 }
+internal data class OrderListTroubleshootingPresentation(
+    val type: OrderListTroubleshootingType,
+    val isExpanded: Boolean = true,
+)
+
+internal enum class OrderListTroubleshootingType {
+    ParsingError,
+    Timeout,
+}
+
+@OptIn(ExperimentalLayoutApi::class)
+@Composable
+private fun OrderListTroubleshooting(
+    presentation: OrderListTroubleshootingPresentation,
+    onExpandedChanged: (Boolean) -> Unit,
+    onTroubleshootingClicked: () -> Unit,
+    onContactSupportClicked: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    Column(
+        modifier = modifier
+            .fillMaxWidth()
+            .testTag(OrderListScreenTestTags.TROUBLESHOOTING),
+    ) {
+        Surface(
+            color = WooTheme.colors.surface.default,
+        ) {
+            Column {
+                Row(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .toggleable(
+                            value = presentation.isExpanded,
+                            role = Role.Button,
+                            onValueChange = onExpandedChanged,
+                        )
+                        .testTag(OrderListScreenTestTags.TROUBLESHOOTING_TOGGLE)
+                        .padding(
+                            horizontal = WooTheme.padding.padding7,
+                            vertical = WooTheme.padding.padding4,
+                        ),
+                    verticalAlignment = Alignment.CenterVertically,
+                ) {
+                    Icon(
+                        imageVector = WooIcons.Regular.CircleInfo,
+                        contentDescription = null,
+                        tint = WooTheme.colors.primary,
+                        modifier = Modifier.size(WooTheme.iconSize.size24),
+                    )
+                    Spacer(modifier = Modifier.width(WooTheme.spacing.space4))
+                    Text(
+                        text = stringResource(presentation.type.titleRes),
+                        modifier = Modifier
+                            .weight(1f)
+                            .semantics { heading() },
+                        color = WooTheme.colors.surface.onDefault,
+                        style = WooTheme.text.titleMedium.strong,
+                    )
+                    Spacer(modifier = Modifier.width(WooTheme.spacing.space3))
+                    Icon(
+                        imageVector = if (presentation.isExpanded) {
+                            WooIcons.Regular.AngleUp
+                        } else {
+                            WooIcons.Regular.AngleDown
+                        },
+                        contentDescription = null,
+                        tint = WooTheme.colors.surface.onVariant,
+                        modifier = Modifier.size(WooTheme.iconSize.size18),
+                    )
+                }
+                AnimatedVisibility(visible = presentation.isExpanded) {
+                    Column(
+                        modifier = Modifier.padding(
+                            start = WooTheme.padding.padding7,
+                            end = WooTheme.padding.padding7,
+                            bottom = WooTheme.padding.padding5,
+                        ),
+                        verticalArrangement = Arrangement.spacedBy(WooTheme.spacing.space4),
+                    ) {
+                        Text(
+                            text = stringResource(presentation.type.messageRes),
+                            color = WooTheme.colors.surface.onVariant,
+                            style = WooTheme.text.bodyMedium.regular,
+                        )
+                        FlowRow(
+                            modifier = Modifier.fillMaxWidth(),
+                            horizontalArrangement = Arrangement.spacedBy(
+                                WooTheme.spacing.space3,
+                                Alignment.End,
+                            ),
+                            verticalArrangement = Arrangement.spacedBy(WooTheme.spacing.space3),
+                        ) {
+                            WooOutlinedButton(
+                                text = stringResource(R.string.error_troubleshooting),
+                                onClick = onTroubleshootingClicked,
+                                size = WooButtonSize.Small,
+                                modifier = Modifier.testTag(
+                                    OrderListScreenTestTags.TROUBLESHOOTING_ACTION
+                                ),
+                            )
+                            WooOutlinedButton(
+                                text = stringResource(R.string.support_contact),
+                                onClick = onContactSupportClicked,
+                                size = WooButtonSize.Small,
+                                modifier = Modifier.testTag(
+                                    OrderListScreenTestTags.CONTACT_SUPPORT_ACTION
+                                ),
+                            )
+                        }
+                    }
+                }
+            }
+        }
+        WooDivider()
+    }
+}
+
+@get:StringRes
+private val OrderListTroubleshootingType.titleRes: Int
+    get() = when (this) {
+        OrderListTroubleshootingType.ParsingError -> R.string.orderlist_parsing_error_title
+        OrderListTroubleshootingType.Timeout -> R.string.orderlist_timeout_error_title
+    }
+
+@get:StringRes
+private val OrderListTroubleshootingType.messageRes: Int
+    get() = when (this) {
+        OrderListTroubleshootingType.ParsingError -> R.string.orderlist_parsing_error_message
+        OrderListTroubleshootingType.Timeout -> R.string.orderlist_timeout_error_message
+    }
+
+@PreviewLightDark
+@Composable
+private fun OrderListBrowsingPreview() {
+    OrderListScreenPreview(
+        state = OrderListScreenState(
+            lastUpdate = "Last updated Jul 28, 10:42 AM",
+        ),
+    )
+}
+
+@Preview(name = "Filtered with JITM", heightDp = 780)
+@Composable
+private fun OrderListFilteredWithJitmPreview() {
+    OrderListScreenPreview(
+        state = OrderListScreenState(
+            filterCount = 3,
+            lastUpdate = "Last updated Jul 28, 10:42 AM",
+        ),
+        jitmContent = {
+            WooNoticeBanner(
+                title = "Grow your business with Woo",
+                description = "A just-in-time message supplied by the Orders host.",
+                tone = WooNoticeBannerTone.Info,
+                modifier = Modifier.padding(
+                    horizontal = WooTheme.padding.padding5,
+                    vertical = WooTheme.padding.padding3,
+                ),
+            )
+        },
+    )
+}
+
+@Preview(name = "Search", heightDp = 700)
+@Composable
+private fun OrderListSearchPreview() {
+    OrderListScreenPreview(
+        state = OrderListScreenState(
+            isSearchActive = true,
+            searchQuery = "#1002",
+            filterCount = 2,
+        ),
+        items = previewOrderListItems.takeLast(1),
+    )
+}
+
+@Preview(name = "Selection over search", heightDp = 700)
+@Composable
+private fun OrderListSelectionPreview() {
+    OrderListScreenPreview(
+        state = OrderListScreenState(
+            isSearchActive = true,
+            searchQuery = "Ada",
+            rowState = OrderListRowState(
+                bulkSelectedOrderIds = setOf(1L),
+                detailHighlightedOrderId = 2L,
+            ),
+        ),
+    )
+}
+
+@PreviewLightDark
+@Composable
+private fun OrderListTroubleshootingPreview() {
+    OrderListScreenPreview(
+        state = OrderListScreenState(
+            troubleshooting = OrderListTroubleshootingPresentation(
+                type = OrderListTroubleshootingType.Timeout,
+            ),
+        ),
+    )
+}
+
+@Preview(name = "Troubleshooting collapsed", heightDp = 700)
+@Composable
+private fun OrderListTroubleshootingCollapsedPreview() {
+    OrderListScreenPreview(
+        state = OrderListScreenState(
+            troubleshooting = OrderListTroubleshootingPresentation(
+                type = OrderListTroubleshootingType.ParsingError,
+                isExpanded = false,
+            ),
+        ),
+    )
+}
+
+@Preview(name = "Narrow large font", widthDp = 320, heightDp = 700, fontScale = 2f)
+@Composable
+private fun OrderListLargeFontPreview() {
+    OrderListScreenPreview(
+        state = OrderListScreenState(
+            filterCount = 12,
+            lastUpdate = "Last updated Jul 28, 10:42 AM",
+        ),
+    )
+}
+
+@Composable
+private fun OrderListScreenPreview(
+    state: OrderListScreenState,
+    items: List<OrderListItemUiModel?> = previewOrderListItems,
+    jitmContent: (@Composable () -> Unit)? = null,
+) {
+    WooDesignSystemThemeWithBackground {
+        OrderListScreen(
+            state = state,
+            orderListContent = { modifier ->
+                OrderListContent(
+                    state = OrderListContentState.Content(),
+                    rowState = state.rowState,
+                    itemCount = items.size,
+                    itemKey = { index -> previewOrderListItemKey(index, items[index]) },
+                    itemAt = items::get,
+                    onOrderTapped = {},
+                    onOrderLongPressed = {},
+                    onOrderSelectionToggled = { true },
+                    onMarkOrderCompleted = {},
+                    onLearnMoreClicked = {},
+                    onShowGuestOrdersClicked = {},
+                    onRetryClicked = {},
+                    modifier = modifier,
+                )
+            },
+            onSearchClicked = {},
+            onSearchQueryChanged = {},
+            onSearchSubmitted = {},
+            onSearchClosed = {},
+            onBarcodeClicked = {},
+            onFiltersClicked = {},
+            onCreateOrderClicked = {},
+            onSelectionCloseClicked = {},
+            onSelectionUpdateStatusClicked = {},
+            onTroubleshootingExpandedChanged = {},
+            onTroubleshootingClicked = {},
+            onContactSupportClicked = {},
+            jitmContent = jitmContent,
+        )
+    }
+}
+
+private fun previewOrderListItemKey(
+    index: Int,
+    item: OrderListItemUiModel?,
+): Any = when (item) {
+    is OrderListItemUiModel.DateSection -> "section-${item.title}"
+    is OrderListItemUiModel.Order -> "order-${item.orderId}"
+    is OrderListItemUiModel.Loading -> "loading-${item.orderId}"
+    null -> "placeholder-$index"
+}
+
+private val previewOrderListItems = listOf(
+    OrderListItemUiModel.DateSection("Today"),
+    previewOrderListOrder(
+        orderId = 1L,
+        number = "#1001",
+        customerName = "Ada Lovelace",
+        total = "\$48.00",
+    ),
+    previewOrderListOrder(
+        orderId = 2L,
+        number = "#1002",
+        customerName = "Grace Hopper",
+        total = "\$86.50",
+        showDivider = false,
+    ),
+)
+
+private fun previewOrderListOrder(
+    orderId: Long,
+    number: String,
+    customerName: String,
+    total: String,
+    showDivider: Boolean = true,
+) = OrderListItemUiModel.Order(
+    orderId = orderId,
+    number = number,
+    customerName = customerName,
+    dateCreated = "Jul 28, 2026 10:30",
+    total = total,
+    badges = listOf(
+        OrderListBadgeUiModel(
+            text = "Processing",
+            containerColorRes = R.color.tag_bg_processing,
+            contentColorRes = R.color.tagView_text,
+        ),
+        OrderListBadgeUiModel(
+            text = "POS",
+            containerColorRes = R.color.tag_bg_pos,
+            contentColorRes = R.color.tag_text_pos,
+        ),
+    ),
+    showDivider = showDivider,
+)
